+++ src/types/index.ts
@@ -3,12 +3,10 @@
 *   Copyright © 2024 NatML Inc. All Rights Reserved.
 */
 
-export * from "./dtype"
 export * from "./environment"
 export * from "./prediction"
 export * from "./predictor"
 export * from "./profile"
 export * from "./tag"
-export * from "./upload"
 export * from "./user"
 export * from "./value"