+++ src/types/value.ts
@@ -3,7 +3,36 @@
 *   Copyright © 2024 NatML Inc. All Rights Reserved.
 */
 
-import { Dtype } from "./dtype"
+/**
+ * Value data type.
+ */
+export type Dtype =
+    "int8"      |
+    "int16"     |
+    "int32"     |
+    "int64"     |
+    "uint8"     |
+    "uint16"    |
+    "uint32"    |
+    "uint64"    |
+    "float16"   |
+    "float32"   |
+    "float64"   |
+    "bool"      |
+    "string"    |
+    "list"      |
+    "dict"      |
+    "image"     |
+    "audio"     |
+    "video"     |
+    "model"     |
+    "binary"    |
+    "null";
+
+/**
+ * Boolean typed array.
+ */
+export class BoolArray extends Uint8Array { }
 
 /**
  * A `TypedArray` instance.
@@ -19,7 +48,8 @@ export type TypedArray =
     Uint8ClampedArray   |
     Uint16Array         |
     Uint32Array         |
-    BigUint64Array;
+    BigUint64Array      |
+    BoolArray;
 
 /**
  * Image.